@@ -1,35 +1,34 @@
 ﻿using System;
 
-namespace ViewStore.Cache
+namespace ViewStore.Cache;
+
+public sealed class DrainStatistics
 {
-    public sealed class DrainStatistics
-    {
-        public TimeSpan Duration { get; }
-        public int AddedOrUpdatedViewCount { get; }
-        public int AddedOrUpdatedRetryCount { get; }
-        public int DeletedViewCount { get; }
-        public int DeletedRetryCount { get; }
+    public TimeSpan Duration { get; }
+    public int AddedOrUpdatedViewCount { get; }
+    public int AddedOrUpdatedRetryCount { get; }
+    public int DeletedViewCount { get; }
+    public int DeletedRetryCount { get; }
         
-        public int Count => AddedOrUpdatedViewCount + DeletedViewCount;
-        public bool HasAnyItems => Count > 0;
+    public int Count => AddedOrUpdatedViewCount + DeletedViewCount;
+    public bool HasAnyItems => Count > 0;
 
-        internal DrainStatistics(
-            TimeSpan duration,
-            int addedOrUpdatedViewCount,
-            int addedOrUpdatedRetryCount,
-            int deletedViewCount,
-            int deletedRetryCount)
-        {
-            Duration = duration;
-            AddedOrUpdatedViewCount = addedOrUpdatedViewCount;
-            AddedOrUpdatedRetryCount = addedOrUpdatedRetryCount;
-            DeletedViewCount = deletedViewCount;
-            DeletedRetryCount = deletedRetryCount;
-        }
+    internal DrainStatistics(
+        TimeSpan duration,
+        int addedOrUpdatedViewCount,
+        int addedOrUpdatedRetryCount,
+        int deletedViewCount,
+        int deletedRetryCount)
+    {
+        Duration = duration;
+        AddedOrUpdatedViewCount = addedOrUpdatedViewCount;
+        AddedOrUpdatedRetryCount = addedOrUpdatedRetryCount;
+        DeletedViewCount = deletedViewCount;
+        DeletedRetryCount = deletedRetryCount;
+    }
 
-        public override string ToString()
-        {
-            return $"{nameof(Duration)}: {Duration}, {nameof(AddedOrUpdatedViewCount)}: {AddedOrUpdatedViewCount}, {nameof(AddedOrUpdatedRetryCount)}: {AddedOrUpdatedRetryCount}, {nameof(DeletedViewCount)}: {DeletedViewCount}, {nameof(DeletedRetryCount)}: {DeletedRetryCount}";
-        }
+    public override string ToString()
+    {
+        return $"{nameof(Duration)}: {Duration}, {nameof(AddedOrUpdatedViewCount)}: {AddedOrUpdatedViewCount}, {nameof(AddedOrUpdatedRetryCount)}: {AddedOrUpdatedRetryCount}, {nameof(DeletedViewCount)}: {DeletedViewCount}, {nameof(DeletedRetryCount)}: {DeletedRetryCount}";
     }
 }
