@@ -1,26 +1,25 @@
 using System;
 
-namespace ViewStore.Cache
+namespace ViewStore.Cache;
+
+public sealed class ThrottleStatistics
 {
-    public sealed class ThrottleStatistics
+    public long CacheCount { get; }
+    public int ThrottleAfterCacheCount { get; }
+    public TimeSpan ThrottlePeriod { get; }
+
+    public ThrottleStatistics(
+        long cacheCount,
+        int throttleAfterCacheCount,
+        TimeSpan throttlePeriod)
     {
-        public long CacheCount { get; }
-        public int ThrottleAfterCacheCount { get; }
-        public TimeSpan ThrottlePeriod { get; }
+        CacheCount = cacheCount;
+        ThrottleAfterCacheCount = throttleAfterCacheCount;
+        ThrottlePeriod = throttlePeriod;
+    }
 
-        public ThrottleStatistics(
-            long cacheCount,
-            int throttleAfterCacheCount,
-            TimeSpan throttlePeriod)
-        {
-            CacheCount = cacheCount;
-            ThrottleAfterCacheCount = throttleAfterCacheCount;
-            ThrottlePeriod = throttlePeriod;
-        }
-
-        public override string ToString()
-        {
-            return $"{nameof(CacheCount)}: {CacheCount}, {nameof(ThrottleAfterCacheCount)}: {ThrottleAfterCacheCount}, {nameof(ThrottlePeriod)}: {ThrottlePeriod}";
-        }
+    public override string ToString()
+    {
+        return $"{nameof(CacheCount)}: {CacheCount}, {nameof(ThrottleAfterCacheCount)}: {ThrottleAfterCacheCount}, {nameof(ThrottlePeriod)}: {ThrottlePeriod}";
     }
 }
