@@ -2,129 +2,128 @@
 using System.Collections.Generic;
 using System.Runtime.Caching;
 
-namespace ViewStore.Cache
+namespace ViewStore.Cache;
+
+public sealed class ViewStoreCacheBuilder
 {
-    public sealed class ViewStoreCacheBuilder
+    private IDatabaseProvider? _databaseProvider;
+    private TimeSpan _cacheDrainPeriod = TimeSpan.FromSeconds(5);
+    private int _cacheDrainBatchSize = 1000;
+    private int _throttleAfterCacheCount = 50000;
+    private readonly List<Action<DrainStatistics>> _cacheDrainedCallbacks = new();
+    private Action<ThrottleStatistics>? _throttlingCallback;
+    private Action<Exception>? _onDrainAttemptFailedCallback;
+    private bool _isBackgroundWorker;
+        
+    private MemoryCache _memoryCache = MemoryCache.Default;
+    private TimeSpan _readCacheExpirationPeriod = TimeSpan.FromSeconds(10);
+
+    public static ViewStoreCacheBuilder New() => new();
+
+    public ViewStoreCacheBuilder WithDatabaseProvider(IDatabaseProvider databaseProvider)
     {
-        private IDatabaseProvider? _databaseProvider;
-        private TimeSpan _cacheDrainPeriod = TimeSpan.FromSeconds(5);
-        private int _cacheDrainBatchSize = 1000;
-        private int _throttleAfterCacheCount = 50000;
-        private readonly List<Action<DrainStatistics>> _cacheDrainedCallbacks = new();
-        private Action<ThrottleStatistics>? _throttlingCallback;
-        private Action<Exception>? _onDrainAttemptFailedCallback;
-        private bool _isBackgroundWorker;
+        _databaseProvider = databaseProvider;
+        return this;
+    }
+
+    public ViewStoreCacheBuilder WithReadMemoryCache(MemoryCache memoryCache)
+    {
+        _memoryCache = memoryCache;
+        return this;
+    }
         
-        private MemoryCache _memoryCache = MemoryCache.Default;
-        private TimeSpan _readCacheExpirationPeriod = TimeSpan.FromSeconds(10);
+    public ViewStoreCacheBuilder WithReadCacheExpirationPeriod(TimeSpan readCacheExpirationPeriod)
+    {
+        _readCacheExpirationPeriod = readCacheExpirationPeriod;
+        return this;
+    }
 
-        public static ViewStoreCacheBuilder New() => new();
+    public ViewStoreCacheBuilder WithCacheDrainPeriod(TimeSpan timeSpan)
+    {
+        _cacheDrainPeriod = timeSpan;
+        return this;
+    }
 
-        public ViewStoreCacheBuilder WithDatabaseProvider(IDatabaseProvider databaseProvider)
+    public ViewStoreCacheBuilder WithCacheDrainBatchSize(int batchSize)
+    {
+        if (batchSize < 0)
         {
-            _databaseProvider = databaseProvider;
-            return this;
+            throw new ArgumentException(nameof(batchSize));
         }
 
-        public ViewStoreCacheBuilder WithReadMemoryCache(MemoryCache memoryCache)
-        {
-            _memoryCache = memoryCache;
-            return this;
-        }
+        _cacheDrainBatchSize = batchSize;
+        return this;
+    }
         
-        public ViewStoreCacheBuilder WithReadCacheExpirationPeriod(TimeSpan readCacheExpirationPeriod)
-        {
-            _readCacheExpirationPeriod = readCacheExpirationPeriod;
-            return this;
-        }
+    public ViewStoreCacheBuilder UseBackgroundWorker()
+    {
+        _isBackgroundWorker = true;
+        return this;
+    }
 
-        public ViewStoreCacheBuilder WithCacheDrainPeriod(TimeSpan timeSpan)
+    public ViewStoreCacheBuilder WithThrottleAfterCacheCount(int throttleAfterCacheCount)
+    {
+        if (throttleAfterCacheCount <= 0)
         {
-            _cacheDrainPeriod = timeSpan;
-            return this;
+            throw new ArgumentException(nameof(throttleAfterCacheCount));
         }
-
-        public ViewStoreCacheBuilder WithCacheDrainBatchSize(int batchSize)
-        {
-            if (batchSize < 0)
-            {
-                throw new ArgumentException(nameof(batchSize));
-            }
-
-            _cacheDrainBatchSize = batchSize;
-            return this;
-        }
-        
-        public ViewStoreCacheBuilder UseBackgroundWorker()
-        {
-            _isBackgroundWorker = true;
-            return this;
-        }
-
-        public ViewStoreCacheBuilder WithThrottleAfterCacheCount(int throttleAfterCacheCount)
-        {
-            if (throttleAfterCacheCount <= 0)
-            {
-                throw new ArgumentException(nameof(throttleAfterCacheCount));
-            }
             
-            _throttleAfterCacheCount = throttleAfterCacheCount;
-            return this;
-        }
+        _throttleAfterCacheCount = throttleAfterCacheCount;
+        return this;
+    }
 
-        public ViewStoreCacheBuilder UseCallbackWhenDrainFinished(Action<DrainStatistics> callback)
-        {
-            _cacheDrainedCallbacks.Add(callback);
-            return this;
-        }
+    public ViewStoreCacheBuilder UseCallbackWhenDrainFinished(Action<DrainStatistics> callback)
+    {
+        _cacheDrainedCallbacks.Add(callback);
+        return this;
+    }
         
-        public ViewStoreCacheBuilder UseCallbackOnThrottling(Action<ThrottleStatistics> throttlingCallback)
-        {
-            _throttlingCallback = throttlingCallback;
-            return this;
-        }
+    public ViewStoreCacheBuilder UseCallbackOnThrottling(Action<ThrottleStatistics> throttlingCallback)
+    {
+        _throttlingCallback = throttlingCallback;
+        return this;
+    }
 
-        public ViewStoreCacheBuilder UseCallbackOnDrainAttemptFailed(Action<Exception> callback)
-        {
-            _onDrainAttemptFailedCallback = callback;
-            return this;
-        }
+    public ViewStoreCacheBuilder UseCallbackOnDrainAttemptFailed(Action<Exception> callback)
+    {
+        _onDrainAttemptFailedCallback = callback;
+        return this;
+    }
 
-        public ViewStoreCache Build()
+    public ViewStoreCache Build()
+    {
+        if (_databaseProvider == null)
         {
-            if (_databaseProvider == null)
-            {
-                throw new ArgumentException(nameof(_databaseProvider));
-            }
+            throw new ArgumentException(nameof(_databaseProvider));
+        }
             
-            var outgoingCache = new OutgoingCache(_throttleAfterCacheCount, _throttlingCallback);
+        var outgoingCache = new OutgoingCache(_throttleAfterCacheCount, _throttlingCallback);
 
-            var automaticCacheDrainer = new AutomaticCacheDrainer(
-                new ManualCacheDrainer(_databaseProvider, outgoingCache, _cacheDrainBatchSize),
-                _cacheDrainPeriod,
-                _isBackgroundWorker);
+        var automaticCacheDrainer = new AutomaticCacheDrainer(
+            new ManualCacheDrainer(_databaseProvider, outgoingCache, _cacheDrainBatchSize),
+            _cacheDrainPeriod,
+            _isBackgroundWorker);
 
-            automaticCacheDrainer.OnDrainFinishedEvent += ds =>
+        automaticCacheDrainer.OnDrainFinishedEvent += ds =>
+        {
+            foreach (var callback in _cacheDrainedCallbacks)
             {
-                foreach (var callback in _cacheDrainedCallbacks)
-                {
-                    callback(ds);
-                }
-            };
-            automaticCacheDrainer.OnSendingExceptionEvent += exception => _onDrainAttemptFailedCallback?.Invoke(exception);
+                callback(ds);
+            }
+        };
+        automaticCacheDrainer.OnSendingExceptionEvent += exception => _onDrainAttemptFailedCallback?.Invoke(exception);
 
-            var readThroughViewStoreCache = new ReadThroughDatabaseProviderCache(
-                _memoryCache,
-                _readCacheExpirationPeriod,
-                _databaseProvider);
+        var readThroughViewStoreCache = new ReadThroughDatabaseProviderCache(
+            _memoryCache,
+            _readCacheExpirationPeriod,
+            _databaseProvider);
 
-            var viewStoreCacheInternal = new ViewStoreCacheInternal(
-                readThroughViewStoreCache,
-                outgoingCache);
+        var viewStoreCacheInternal = new ViewStoreCacheInternal(
+            readThroughViewStoreCache,
+            outgoingCache);
 
-            return new ViewStoreCache(
-                viewStoreCacheInternal,
-                automaticCacheDrainer);
-        }
+        return new ViewStoreCache(
+            viewStoreCacheInternal,
+            automaticCacheDrainer);
     }
 }
